fix: rename lib.js to lib.web.js
This fixes TypeScript support for NodeNext/Node16/Bundler
moduleResolution.

The issue was that the ./dist/lib.js had a higher precedence than
./dist/lib/index.d.ts when TypeScript tried to resolve "./lib"
imports/exports.

Instead of renaming the whole lib/ folder I decided to add a infix to
the lib.js artifact. But to solve the issue you could've renamed the
directory instead.